+++ DataModels/Entities/LogBookInstrumentApproach.cs
@@ -11,7 +11,7 @@ namespace DataModels.Entities
         public long Id { get; set; }
         public long LogBookInstrumentId { get; set; }
         public string Airport { get; set; }
-        public long InstrumentApproachId { get; set; }
+        public short InstrumentApproachId { get; set; }
         public string Runway { get; set; }
         public bool IsCircleToLand { get; set; }
         public string Comments { get; set; }
